Updated styles. Also, ran into some limitations of iframe while trying to add a fallback image incase the banner video background is unreachable. The goal was to show the video background if video hosting service is down, to ensure thee website banner still shows an image (stored within the project folder) instead of showing a gray background with the broken embed icon (paper with folder corner and sad face). Turns out iframe doesn't allow JavaScript inside the iframe to fail gracefully or emit an error when the iframe is unreachable. Tried a workaround: insert only the video iframe if the browser onlinem which means the HTML doesnt have the video iframe embedded by default and will rely on the custom JS code to embed the iframe video ONLY WHEN it detects that the browser is online.
To test this, I ran the website locally via web server, and then to simulate lost of connection to the video hosting service, I disconnected my PC to the computer. After adding the custom JS code and diconnecting internet, the banner still showed the broken embed icon, so apparently the JS code still not working. Since i'm tired for tonight, will just let the original iframe, and accept the limitations of iframe. Might improve this in the future but its a wrap for tonight